consolidate center_freq() getter for GSLC and GCOV
