+++ src/nisarqa/input_product_readers/gslc_reader.py
@@ -18,34 +18,5 @@ class GSLC(SLCProduct, NonInsarGeoProduct):
     def product_type(self) -> str:
         return "GSLC"
 
-    def center_freq(self, freq: str) -> float:
-        """
-        The processed center frequency for input product's Frequency `freq`.
-
-        For GSLC products, this is read from the `centerFrequency` dataset
-        in the grids frequency group.
-
-        Parameters
-        ----------
-        freq : str
-            Must be either "A" or "B".
-
-        Returns
-        -------
-        center_freq : float
-            The processed center frequency for input product's Frequency `freq`,
-            in hertz.
-        """
-
-        freq_group = self.get_freq_path(freq=freq)
-
-        with h5py.File(self.filepath) as f:
-            try:
-                center_freq = f[freq_group]["centerFrequency"][()]
-            except KeyError as e:
-                raise nisarqa.DatasetNotFoundError from e
-
-        return center_freq
-
 
 __all__ = nisarqa.get_all(__name__, objects_to_skip)
